Grantham deviation

This function calculates the Grantham deviation (\(\mathrm{gd}\)):

$$\mathrm{gd} = \rho \left((\alpha\ \mathrm{dev}^2(c_x, c_{min}, c_{max}) + \beta\ \mathrm{dev}^2(p_x, p_{min}, p_{max}) + \gamma\ \mathrm{dev}^2(v_x, v_{min}, v_{max})\right)^\frac{1}{2}$$

where \(c_x\) is the value for composition \(c\) of amino acid \(x\), i.e. the atomic weight ratio of hetero (noncarbon) elements in end groups or rings to carbons in the side chain; \(p_x\) is the value for polarity \(p\) of amino acid \(x\); and, \(v_x\) is the value for molecular volume \(v\) of amino acid \(x\).

\(c_x\), \(p_x\) and \(v_x\) are looked up in grantham::amino_acids_properties based on the amino acid identities passed in x. The function \(\mathrm{dev}\) is implemented in dev(). Remaining variables in the equation are arguments to gd() and hence are explained below in the Arguments section.

Usage

gd(
  x,
  c_min,
  c_max,
  p_min,
  p_max,
  v_min,
  v_max,
  alpha = 1.833,
  beta = 0.1018,
  gamma = 0.000399,
  rho = 50.723
)

Arguments

x

A character vector of one-letter amino acid codes, indicating missense substitutions.

c_min

Amino acid composition, minimum value.

c_max

Amino acid, composition, maximum value.

p_min

Amino acid polarity, minimum value.

p_max

Amino acid polarity, maximum value.

v_min

Amino acid molecular volume, maximum value.

v_max

Amino acid molecular volume, maximum value.

alpha

The constant \(\alpha\) in Grantham's equation. It is the square inverse of the mean of the composition property.

beta

The constant \(\beta\) in Grantham's equation. It is the square inverse of the mean of the polarity property.

gamma

The constant \(\gamma\) in Grantham's equation. It is the square inverse of the mean of the molecular volume property.

rho

Grantham's distances reported in Table 2, Science (1974). 185(4154): 862--4 by R. Grantham, are scaled by a factor (here named \(\rho\)) such that the mean value of all distances are 100. The rho parameter allows this factor \(\rho\) to be changed. By default \(\rho=50.723\), the same value used by Grantham. This value is originally mentioned in the caption of Table 2 of the aforementioned paper.

Value

A numeric vector of Grantham deviations. Each deviation corresponds to one of the amino acids indicated in x.

See also

Examples

gd('S', c_min = 0.39, c_max = 0.74, p_min =4.9, p_max =8.6, v_min = 3, v_max = 32.5)
#> [1] 47.69654