Read in AGVGD Web Results

This function imports into R the results generated by the AGVGD Web application http://agvgd.hci.utah.edu/.

Usage

read_agvgdweb_results(file = stop("`file` is missing"), alignment = NULL)

Arguments

file

A file path to the results output by the AGVGD Web app.

alignment

A character matrix or an alignment object obtained with read_alignment(). Rows are expected to be sequences of single characters (protein residues), and columns the alignment positions. The first row must be the reference sequence, i.e. the sequence whose substitutions will be evaluated against. This parameter can be left NULL. If supplied the column poi in the output will be filled in (default is to be NA).

Value

A tibble of seven columns:

res

Position of the amino acid residue in the reference protein (first sequence in the alignment). This position corresponds to poi minus the gaps in the alignment.

poi

Position of interest, i.e. the alignment position at which the amino acid substitution is being assessed. Because this information is not provided by AGVGD Web app this column is always NA; we keep it though for coherence with the output of agvgd().

ref

Reference amino acid, i.e. the amino acid in the first sequence of the alignment, at the position of interest.

sub

Amino acid substitution being assessed.

gv

Grantham variation score.

gd

Grantham difference score.

prediction

Predicted effect of the amino acid substitution. This is classed as C0, C15, C25, C35, C45, C55, or C65, with C65 most likely to interfere with function and C0 least likely.