Read a protein sequence multiple alignment

Reads a protein sequence multiple alignment (PSMA) from either a set of pre-bundled alignments, by gene name, or from a Multi-FASTA file.

Usage

read_alignment(
  gene = c("ATM", "BRCA1", "BRCA2", "CHEK2", "MRE11", "MSH6", "NBN", "PALB2", "PMS2",
    "RAD50", "RAD51", "XRCC2"),
  file = NULL
)

Arguments

gene

The gene name for which an alignment is provided with this package. Use the function alignment_file() to list the pre-bundled alignments.

file

The path to a Multi-FASTA file. If this argument is given, it takes precedence over the gene parameter.

Value

An alignment object; essentially, a character matrix, whose elements are protein residues in one-letter notation. Rows are sequences and columns are alignment positions.

Examples

# Read in the alignment for the gene XRCC2
read_alignment('XRCC2')
#> Hsap_XRCC2   1 MCSAFHRAESGTELLARLEGRSSLKEIEPNLFADEDS--PVHGDILEFHG
#> Mmul_XRCC2   1 MCSDFHRAESGTELLARLEGRSSLKEIEPNLFADEDS--PVHGDILEFHG
#> Mmus_XRCC2   1 MCSDFRRAESGTELLARLEGRSSLKELEPNLFADEDS--PVHGDIFEFHG
#> Cfam_XRCC2   1 MCSDFHRAESGTELLARLEGRSSLKEIEPYLFTDEVS--SVHGDILEFHG
#> Lafr_XRCC2   1 MCSDFHRAESGTELLARLEGRSSLKVIEPYLFADEES--PVHGDILEFHG
#> Mdom_XRCC2   1 MSGDFRRAESGTELLARLEGRSSLKDIEPFLFADEGS--PIHGDILEFHG
#> Oana_XRCC2   1 MSGHFRRAESGTELLARLEGRSSLKTLEPFLFADEGF--PIHGDILEFHG
#> Ggal_XRCC2   1 MGDAFRRAESGTQLLARLEGRSSLKNLEPNLFAEEGS--PVHGDVIEFHG
#> Acar_XRCC2   1 MTGRFGEAESGAQLLARLEGRGSLKDLEPCLFAEEGY--PIPGDIIECYG
#> Xtro_XRCC2   1 MSDGSRQAESGTQLLARLEGRASLSNLEPLLFADEGC--PVHGEITEFYG
#> Drer_XRCC2   1 MTARVRMAENGAQLVSRLEGRQSLKDIEPNIFPADGG--PGQGDVVEFHG
#> Bflo_XRCC2   1 MXXXXXXXXXXXXLLARLGSRPSLVQLETALFRADMG--PKSGDAIELYG
#> Spur_XRCC2   1 MXXXXXXXXXXXXLFARLGEKPSLARLNPKLIPPGLE--PRPGDVVEIYG
#> Nvec_XRCC2   1 MXXXXXXXXXXXKLFSRLGSKQSLDGLDKKLFVDIPD-GIKAGDVVEFYG
#> Tadh_XRCC2   1 M-----ASESAAKLFARLGSRQTVIGMEDRLFSKLQFNGLTCGDVVEFYG 
#> 
#> Hsap_XRCC2  51 PEGTGKTEMLYHLTARCILPKS-----EGGLEVEVLFIDTDYHFDMLRLV
#> Mmul_XRCC2  51 SEGTGKTEMLYHLTARCILPKS-----EGGLEVEVLFIDTDYHFDMLRLV
#> Mmus_XRCC2  51 PEGTGKTEMLYHLTARCILPKS-----EGGLQIEVLFIDTDYHFDMLRLV
#> Cfam_XRCC2  51 PEGTGKTEMLYHLTARCILPKS-----EGGLEVEVLFIDTDYHFDMLRLV
#> Lafr_XRCC2  51 PEGTGKTEMLYHLTARCILPKS-----EGGLEIEVLFIDTDHHFDMLRLV
#> Mdom_XRCC2  51 PEGTGKTEMLYHLTARCILPKS-----EGGLEVEVLFIDTDYHFDMLRLV
#> Oana_XRCC2  51 QEGTGKTEMLYHLVARCILPRS-----EGGLEEEVLFVDTDYHFDMLRLI
#> Ggal_XRCC2  51 PEGTGKTEMLYHLIARCIIPKS-----GGGLEVEVMFIDTDYHFDMLRLV
#> Acar_XRCC2  51 PEGTGKTEMFYHLIARCILPKS-----RGGLEVGLLFIDTDFHFDMLRLV
#> Xtro_XRCC2  51 PEGTGKTEMLCHLISRCILPKS-----DGGLQVEVIYIDTDYHFDMLRLV
#> Drer_XRCC2  51 MEGSGKTETLYHLITRCITPTH-----SGGLEVGVVFIDTDYHFDMLRFV
#> Bflo_XRCC2  51 PEGTAKTEMLLHLTARCILPAS-----VGGLEAGVVFIDNDYHFDILRLV
#> Spur_XRCC2  51 NSGSGKTELLLNLAAMCILPERWKTIDIGGLGTSVVFIDTDHQFSMLRLF
#> Nvec_XRCC2  51 KEGCGKTEMLLHLAANCIMPRSWHELYLGGKGVSVIFIDTDYHFQILRLI
#> Tadh_XRCC2  51 SEGCGKTEMLLHLMIKCIMPDNFRGIAMNGRSMSVVYIDCDYHFNLLRLM 
#> 
#> Hsap_XRCC2 101 TILEHRL------SQSS--------EEIIKYCLGRFFLVYCSSSTHLLLT
#> Mmul_XRCC2 101 TVLEHRL------SQSS--------EEIIKYCLGRFFLVYCSSSTHLLLT
#> Mmus_XRCC2 101 TVLEHRL------SQSS--------EEAMKLCLARLFLAYCSSSMQLLLT
#> Cfam_XRCC2 101 TILEHRL------SQSS--------EEMVKHCLGRLFLVNCNSSTQLLLT
#> Lafr_XRCC2 101 TILEHRL------SQSS--------EEIIKSCLGRFFLVYCSSSSQLLLT
#> Mdom_XRCC2 101 TILERRL------SQST--------EDIIKHCLGRFFLVNCSSSNQLLIT
#> Oana_XRCC2 101 TILEHRL------SQSS--------EEAIKLCLGRLFLVYCSSSVQLLLT
#> Ggal_XRCC2 101 TILEHRL------EQST--------EEMMKRCLGRLFLVNCNSSTQLLLT
#> Acar_XRCC2 101 TILEHRS------SQGT--------EDMIKQCLGRFFLVNCSSSSQLLLT
#> Xtro_XRCC2 101 TILEHRL------AQNT--------EEAVKQCLGRFFLLYCNSSVQLLLT
#> Drer_XRCC2 101 SILEGRL------AEDSKTGSENEAEETVRSCLCRLSVVHCNSSVQLLLT
#> Bflo_XRCC2 101 TVLEGRL------DTTD--------EDRMKQCLGRLYIVRCNSSEQLVIT
#> Spur_XRCC2 101 ALLERKV------AEAIDNRTKRK-ETFLKACLKKLYMVKIATSNQLVIT
#> Nvec_XRCC2 101 AIMEYRT------AESE---------TLIKQCLTRLFIVRCNSSVELLAT
#> Tadh_XRCC2 101 SILEQRYCKACQGSNATMVKHS---EEFIRKCLERFYIIRCDTIHQLITS 
#> 
#> Hsap_XRCC2 151 LYSLESMFCSHPSLCLLILDSLSAFYWIDRVNGGESVNLQESTLRKCSQC
#> Mmul_XRCC2 151 LYSLESMFCSHPSLCLLILDSLSAFYWIDRVNGGESVNLQESTLRKCSQC
#> Mmus_XRCC2 151 LHSLEALLCSRPSLCLLIVDSLSSFYWIDRVSGGESVALQESTLQKCSQL
#> Cfam_XRCC2 151 LYSLETVVCSHPSLCLLILDSLSAFYWIDRVNGGESVNLQEATLKKCAQF
#> Lafr_XRCC2 151 LYSLESMFCSHPSLCLLILDSLSAFYWIDRANGGESVNLQESTLKKCSQF
#> Mdom_XRCC2 151 LYSLETMFCSHPSLCLFILDSLSAFYWIDRVNGGESLTLQEINLKKCSKF
#> Oana_XRCC2 151 LHSLETMFCSRPSLSLLMVDSLSAFYWIDRANGGESLTQQEATLRKCTRL
#> Ggal_XRCC2 151 LYSLENMFCTHPSLCLLILDSISAFYWIDRSNGGESLNSQEMNLKKCANF
#> Acar_XRCC2 151 LYSLENMFCSHPSLCLLIIDSISAFYWIDRVNGGESISLQEANLRRCAQF
#> Xtro_XRCC2 151 LYSLENMFCSHPSLCLLIIDSISAFYWIDRNNGGETFAKQETNLRKCTEL
#> Drer_XRCC2 151 LHYLENTFSSQPTLGLLVIDSISAFYWTDRFNGGESASCQEANLRKCAEL
#> Bflo_XRCC2 151 LHSLEHIIASSSEVALLIVDSISAFYWLDR-STDDSMSGQELNQRRCVDI
#> Spur_XRCC2 151 LHSLESLLASQCDISVLMMDSVSAFYWVDRMKGDG-AHRQGVNQKLAFGA
#> Nvec_XRCC2 151 LLSMEQLIICKPEICVMMIDSLSAFYWVDRSSGGESLQDQQENIRKTTSV
#> Tadh_XRCC2 151 LHLLEYSISSNPDIGIMLIDGIGSFYWQDKFSSSS-----GVDQ--LCKI 
#> 
#> Hsap_XRCC2 201 LEKLVNDYRLVLFATTQTIMQKAS------------SSSEEPSHASRRLC
#> Mmul_XRCC2 201 LEKLVNDYRLVLFATTQTIMQKAL------------NSSEEPSPASRRLR
#> Mmus_XRCC2 201 LERLVTEYRLLLFATTQSLMQKGS------------DSADGPSSS-KHPC
#> Cfam_XRCC2 201 LEKLVNEYRLVLFATTQSIMQKTS------------NWTEGPSSAFNHPK
#> Lafr_XRCC2 201 LERLVNEYRLILFASTQSIMQKPS------------NSTEGPSSAFKQPS
#> Mdom_XRCC2 201 LEKLVKEYHLVLFATTQTIMQKNS------------NSTERSSSL-KLPC
#> Oana_XRCC2 201 LEKLVKEYHLVLLATTQAIMQRSS------------KASENSASA----W
#> Ggal_XRCC2 201 LEKLVREHHLVLFATTQSIMQKST------------NSAEGF-FPLKLQS
#> Acar_XRCC2 201 LEKLVREHHLVLFATTQAIMQKSL------------NAIE---SSRKRNS
#> Xtro_XRCC2 201 LHKLLKEYQLVLFASTQAIMQKSP------------NEAGEGPSRSGKQN
#> Drer_XRCC2 201 LDRLRRNYGIVIFATTHAIMRNFGSDLG--VS-DVHGSSSSSSSRRWRSA
#> Bflo_XRCC2 201 LSRYLSDYGIVLIATKQALFGHKSRKNQ--NE-D-----TTLSPKLEKTK
#> Spur_XRCC2 201 LSRLVEDYHLVLFASKAALVTKQPQN-EFSLRLDSTGETDHSNRTSTTSV
#> Nvec_XRCC2 201 LSRFSRENHLVIFTTVHAIFG---NN----TK----------E--M----
#> Tadh_XRCC2 201 VKYLCDEHNLIVLATKSAIRKQFENSRL--AN-K----SRLRNDIASNYN 
#> 
#> Hsap_XRCC2 251 DVDIDY-RPYLCKAWQQLVKHRMFFSKQDDS----QS-SNQFSLVSRCLK
#> Mmul_XRCC2 251 DVDVDY-RPYLCKAWQQLVKHRIFFSKQDDS----QS-SNQFSLVSRCLK
#> Mmus_XRCC2 251 DGDMGY-RAYLCKAWQRVVKHRVIFSRDDEA----K--SSRFSLVSRHLK
#> Cfam_XRCC2 251 EADADY-RPYLCKEWQQVVKHRIFFSKQEDF----K---TQFSLVSRHLK
#> Lafr_XRCC2 251 NEDIDY-RPYLCKAWQQMVKHRIFFSKQDDS----KR-NNQFSLVSRHLK
#> Mdom_XRCC2 251 EVDIDY-RPYLCKSWQQMVNHRIFFSRNSES----S---NQMSVVSYHLK
#> Oana_XRCC2 251 EGDGDY-RPYLCKSWQQLVNHRLFFSKQDNG----EDPKQMFSFTSCHLK
#> Ggal_XRCC2 251 EIDADY-RPYLCKSWQQMVTHRIFFSKQFNS----GN-STGFTLVSCHLK
#> Acar_XRCC2 251 DGDVDY-RPYLCKSWQQMITHRIFFSKQCNP----DN-TQSFSITACHIR
#> Xtro_XRCC2 251 SSSMDY-KPYLCKLWQQGATHRVLFSKELRN----N--EQIYSITSCHLK
#> Drer_XRCC2 251 DCASDFDRPYLCRAWQRIVTHRVLFTKSHAP----KDHKQILSTACTSIL
#> Bflo_XRCC2 251 VENVEH-YEYMCHAWQNLIKYRYVFSRATKKDISIEENGKDISSFSATMI
#> Spur_XRCC2 251 KLSTDH-HEFMSQEWTKLVTHRMILERHDHMTS--DGPNSSYLSVLKHKA
#> Nvec_XRCC2 251 ---MRN-QDYLCKAWQQSVKYRYMFTKQTEYDGKASQFCSVYV-VQRTSP
#> Tadh_XRCC2 251 STHSKH-MEYMPNVWRKLVKYRYILSKLDSELSGANSYSATYS-VVLEHP 
#> 
#> Hsap_XRCC2 301 SNSLKKHFFIIGESGVEFC-------
#> Mmul_XRCC2 301 SNSFKKHFFIIGESGVEFC-------
#> Mmus_XRCC2 301 SNSLKKHSFMVRESGVEFC-------
#> Cfam_XRCC2 301 SNSLKKHTFVIGENGIEFC-------
#> Lafr_XRCC2 301 SNSLKKHFFIIGESGVEFC-------
#> Mdom_XRCC2 301 SNNLIKRLFSIRESGVHFC-------
#> Oana_XRCC2 301 SNRFAKRFFSIGEGGVQF--------
#> Ggal_XRCC2 301 KKHVAKRSFSIAECGVQFFQ------
#> Acar_XRCC2 301 NNSVIKRSFSILENGVQF--------
#> Xtro_XRCC2 301 TRNGVKRSFRIAESGVQFL-------
#> Drer_XRCC2 301 TKGVKKCSFCVVEDGIKFICDK----
#> Bflo_XRCC2 301 KPVEHKCEFTVADRGISFMGSLYNQC
#> Spur_XRCC2 301 SGNMYSCQFIIDQQGIRV--------
#> Nvec_XRCC2 301 KTESKSNRFLIEEHGVVFIS------
#> Tadh_XRCC2 301 SSIESAERFIVSEEGVVFI-------

# Also read in the alignment for the gene XRCC2, but now by specifying
# directly the path to the file.
path <- system.file("extdata", alignment_file("XRCC2"), package = "agvgd")
read_alignment(file = path)
#> Hsap_XRCC2   1 MCSAFHRAESGTELLARLEGRSSLKEIEPNLFADEDS--PVHGDILEFHG
#> Mmul_XRCC2   1 MCSDFHRAESGTELLARLEGRSSLKEIEPNLFADEDS--PVHGDILEFHG
#> Mmus_XRCC2   1 MCSDFRRAESGTELLARLEGRSSLKELEPNLFADEDS--PVHGDIFEFHG
#> Cfam_XRCC2   1 MCSDFHRAESGTELLARLEGRSSLKEIEPYLFTDEVS--SVHGDILEFHG
#> Lafr_XRCC2   1 MCSDFHRAESGTELLARLEGRSSLKVIEPYLFADEES--PVHGDILEFHG
#> Mdom_XRCC2   1 MSGDFRRAESGTELLARLEGRSSLKDIEPFLFADEGS--PIHGDILEFHG
#> Oana_XRCC2   1 MSGHFRRAESGTELLARLEGRSSLKTLEPFLFADEGF--PIHGDILEFHG
#> Ggal_XRCC2   1 MGDAFRRAESGTQLLARLEGRSSLKNLEPNLFAEEGS--PVHGDVIEFHG
#> Acar_XRCC2   1 MTGRFGEAESGAQLLARLEGRGSLKDLEPCLFAEEGY--PIPGDIIECYG
#> Xtro_XRCC2   1 MSDGSRQAESGTQLLARLEGRASLSNLEPLLFADEGC--PVHGEITEFYG
#> Drer_XRCC2   1 MTARVRMAENGAQLVSRLEGRQSLKDIEPNIFPADGG--PGQGDVVEFHG
#> Bflo_XRCC2   1 MXXXXXXXXXXXXLLARLGSRPSLVQLETALFRADMG--PKSGDAIELYG
#> Spur_XRCC2   1 MXXXXXXXXXXXXLFARLGEKPSLARLNPKLIPPGLE--PRPGDVVEIYG
#> Nvec_XRCC2   1 MXXXXXXXXXXXKLFSRLGSKQSLDGLDKKLFVDIPD-GIKAGDVVEFYG
#> Tadh_XRCC2   1 M-----ASESAAKLFARLGSRQTVIGMEDRLFSKLQFNGLTCGDVVEFYG 
#> 
#> Hsap_XRCC2  51 PEGTGKTEMLYHLTARCILPKS-----EGGLEVEVLFIDTDYHFDMLRLV
#> Mmul_XRCC2  51 SEGTGKTEMLYHLTARCILPKS-----EGGLEVEVLFIDTDYHFDMLRLV
#> Mmus_XRCC2  51 PEGTGKTEMLYHLTARCILPKS-----EGGLQIEVLFIDTDYHFDMLRLV
#> Cfam_XRCC2  51 PEGTGKTEMLYHLTARCILPKS-----EGGLEVEVLFIDTDYHFDMLRLV
#> Lafr_XRCC2  51 PEGTGKTEMLYHLTARCILPKS-----EGGLEIEVLFIDTDHHFDMLRLV
#> Mdom_XRCC2  51 PEGTGKTEMLYHLTARCILPKS-----EGGLEVEVLFIDTDYHFDMLRLV
#> Oana_XRCC2  51 QEGTGKTEMLYHLVARCILPRS-----EGGLEEEVLFVDTDYHFDMLRLI
#> Ggal_XRCC2  51 PEGTGKTEMLYHLIARCIIPKS-----GGGLEVEVMFIDTDYHFDMLRLV
#> Acar_XRCC2  51 PEGTGKTEMFYHLIARCILPKS-----RGGLEVGLLFIDTDFHFDMLRLV
#> Xtro_XRCC2  51 PEGTGKTEMLCHLISRCILPKS-----DGGLQVEVIYIDTDYHFDMLRLV
#> Drer_XRCC2  51 MEGSGKTETLYHLITRCITPTH-----SGGLEVGVVFIDTDYHFDMLRFV
#> Bflo_XRCC2  51 PEGTAKTEMLLHLTARCILPAS-----VGGLEAGVVFIDNDYHFDILRLV
#> Spur_XRCC2  51 NSGSGKTELLLNLAAMCILPERWKTIDIGGLGTSVVFIDTDHQFSMLRLF
#> Nvec_XRCC2  51 KEGCGKTEMLLHLAANCIMPRSWHELYLGGKGVSVIFIDTDYHFQILRLI
#> Tadh_XRCC2  51 SEGCGKTEMLLHLMIKCIMPDNFRGIAMNGRSMSVVYIDCDYHFNLLRLM 
#> 
#> Hsap_XRCC2 101 TILEHRL------SQSS--------EEIIKYCLGRFFLVYCSSSTHLLLT
#> Mmul_XRCC2 101 TVLEHRL------SQSS--------EEIIKYCLGRFFLVYCSSSTHLLLT
#> Mmus_XRCC2 101 TVLEHRL------SQSS--------EEAMKLCLARLFLAYCSSSMQLLLT
#> Cfam_XRCC2 101 TILEHRL------SQSS--------EEMVKHCLGRLFLVNCNSSTQLLLT
#> Lafr_XRCC2 101 TILEHRL------SQSS--------EEIIKSCLGRFFLVYCSSSSQLLLT
#> Mdom_XRCC2 101 TILERRL------SQST--------EDIIKHCLGRFFLVNCSSSNQLLIT
#> Oana_XRCC2 101 TILEHRL------SQSS--------EEAIKLCLGRLFLVYCSSSVQLLLT
#> Ggal_XRCC2 101 TILEHRL------EQST--------EEMMKRCLGRLFLVNCNSSTQLLLT
#> Acar_XRCC2 101 TILEHRS------SQGT--------EDMIKQCLGRFFLVNCSSSSQLLLT
#> Xtro_XRCC2 101 TILEHRL------AQNT--------EEAVKQCLGRFFLLYCNSSVQLLLT
#> Drer_XRCC2 101 SILEGRL------AEDSKTGSENEAEETVRSCLCRLSVVHCNSSVQLLLT
#> Bflo_XRCC2 101 TVLEGRL------DTTD--------EDRMKQCLGRLYIVRCNSSEQLVIT
#> Spur_XRCC2 101 ALLERKV------AEAIDNRTKRK-ETFLKACLKKLYMVKIATSNQLVIT
#> Nvec_XRCC2 101 AIMEYRT------AESE---------TLIKQCLTRLFIVRCNSSVELLAT
#> Tadh_XRCC2 101 SILEQRYCKACQGSNATMVKHS---EEFIRKCLERFYIIRCDTIHQLITS 
#> 
#> Hsap_XRCC2 151 LYSLESMFCSHPSLCLLILDSLSAFYWIDRVNGGESVNLQESTLRKCSQC
#> Mmul_XRCC2 151 LYSLESMFCSHPSLCLLILDSLSAFYWIDRVNGGESVNLQESTLRKCSQC
#> Mmus_XRCC2 151 LHSLEALLCSRPSLCLLIVDSLSSFYWIDRVSGGESVALQESTLQKCSQL
#> Cfam_XRCC2 151 LYSLETVVCSHPSLCLLILDSLSAFYWIDRVNGGESVNLQEATLKKCAQF
#> Lafr_XRCC2 151 LYSLESMFCSHPSLCLLILDSLSAFYWIDRANGGESVNLQESTLKKCSQF
#> Mdom_XRCC2 151 LYSLETMFCSHPSLCLFILDSLSAFYWIDRVNGGESLTLQEINLKKCSKF
#> Oana_XRCC2 151 LHSLETMFCSRPSLSLLMVDSLSAFYWIDRANGGESLTQQEATLRKCTRL
#> Ggal_XRCC2 151 LYSLENMFCTHPSLCLLILDSISAFYWIDRSNGGESLNSQEMNLKKCANF
#> Acar_XRCC2 151 LYSLENMFCSHPSLCLLIIDSISAFYWIDRVNGGESISLQEANLRRCAQF
#> Xtro_XRCC2 151 LYSLENMFCSHPSLCLLIIDSISAFYWIDRNNGGETFAKQETNLRKCTEL
#> Drer_XRCC2 151 LHYLENTFSSQPTLGLLVIDSISAFYWTDRFNGGESASCQEANLRKCAEL
#> Bflo_XRCC2 151 LHSLEHIIASSSEVALLIVDSISAFYWLDR-STDDSMSGQELNQRRCVDI
#> Spur_XRCC2 151 LHSLESLLASQCDISVLMMDSVSAFYWVDRMKGDG-AHRQGVNQKLAFGA
#> Nvec_XRCC2 151 LLSMEQLIICKPEICVMMIDSLSAFYWVDRSSGGESLQDQQENIRKTTSV
#> Tadh_XRCC2 151 LHLLEYSISSNPDIGIMLIDGIGSFYWQDKFSSSS-----GVDQ--LCKI 
#> 
#> Hsap_XRCC2 201 LEKLVNDYRLVLFATTQTIMQKAS------------SSSEEPSHASRRLC
#> Mmul_XRCC2 201 LEKLVNDYRLVLFATTQTIMQKAL------------NSSEEPSPASRRLR
#> Mmus_XRCC2 201 LERLVTEYRLLLFATTQSLMQKGS------------DSADGPSSS-KHPC
#> Cfam_XRCC2 201 LEKLVNEYRLVLFATTQSIMQKTS------------NWTEGPSSAFNHPK
#> Lafr_XRCC2 201 LERLVNEYRLILFASTQSIMQKPS------------NSTEGPSSAFKQPS
#> Mdom_XRCC2 201 LEKLVKEYHLVLFATTQTIMQKNS------------NSTERSSSL-KLPC
#> Oana_XRCC2 201 LEKLVKEYHLVLLATTQAIMQRSS------------KASENSASA----W
#> Ggal_XRCC2 201 LEKLVREHHLVLFATTQSIMQKST------------NSAEGF-FPLKLQS
#> Acar_XRCC2 201 LEKLVREHHLVLFATTQAIMQKSL------------NAIE---SSRKRNS
#> Xtro_XRCC2 201 LHKLLKEYQLVLFASTQAIMQKSP------------NEAGEGPSRSGKQN
#> Drer_XRCC2 201 LDRLRRNYGIVIFATTHAIMRNFGSDLG--VS-DVHGSSSSSSSRRWRSA
#> Bflo_XRCC2 201 LSRYLSDYGIVLIATKQALFGHKSRKNQ--NE-D-----TTLSPKLEKTK
#> Spur_XRCC2 201 LSRLVEDYHLVLFASKAALVTKQPQN-EFSLRLDSTGETDHSNRTSTTSV
#> Nvec_XRCC2 201 LSRFSRENHLVIFTTVHAIFG---NN----TK----------E--M----
#> Tadh_XRCC2 201 VKYLCDEHNLIVLATKSAIRKQFENSRL--AN-K----SRLRNDIASNYN 
#> 
#> Hsap_XRCC2 251 DVDIDY-RPYLCKAWQQLVKHRMFFSKQDDS----QS-SNQFSLVSRCLK
#> Mmul_XRCC2 251 DVDVDY-RPYLCKAWQQLVKHRIFFSKQDDS----QS-SNQFSLVSRCLK
#> Mmus_XRCC2 251 DGDMGY-RAYLCKAWQRVVKHRVIFSRDDEA----K--SSRFSLVSRHLK
#> Cfam_XRCC2 251 EADADY-RPYLCKEWQQVVKHRIFFSKQEDF----K---TQFSLVSRHLK
#> Lafr_XRCC2 251 NEDIDY-RPYLCKAWQQMVKHRIFFSKQDDS----KR-NNQFSLVSRHLK
#> Mdom_XRCC2 251 EVDIDY-RPYLCKSWQQMVNHRIFFSRNSES----S---NQMSVVSYHLK
#> Oana_XRCC2 251 EGDGDY-RPYLCKSWQQLVNHRLFFSKQDNG----EDPKQMFSFTSCHLK
#> Ggal_XRCC2 251 EIDADY-RPYLCKSWQQMVTHRIFFSKQFNS----GN-STGFTLVSCHLK
#> Acar_XRCC2 251 DGDVDY-RPYLCKSWQQMITHRIFFSKQCNP----DN-TQSFSITACHIR
#> Xtro_XRCC2 251 SSSMDY-KPYLCKLWQQGATHRVLFSKELRN----N--EQIYSITSCHLK
#> Drer_XRCC2 251 DCASDFDRPYLCRAWQRIVTHRVLFTKSHAP----KDHKQILSTACTSIL
#> Bflo_XRCC2 251 VENVEH-YEYMCHAWQNLIKYRYVFSRATKKDISIEENGKDISSFSATMI
#> Spur_XRCC2 251 KLSTDH-HEFMSQEWTKLVTHRMILERHDHMTS--DGPNSSYLSVLKHKA
#> Nvec_XRCC2 251 ---MRN-QDYLCKAWQQSVKYRYMFTKQTEYDGKASQFCSVYV-VQRTSP
#> Tadh_XRCC2 251 STHSKH-MEYMPNVWRKLVKYRYILSKLDSELSGANSYSATYS-VVLEHP 
#> 
#> Hsap_XRCC2 301 SNSLKKHFFIIGESGVEFC-------
#> Mmul_XRCC2 301 SNSFKKHFFIIGESGVEFC-------
#> Mmus_XRCC2 301 SNSLKKHSFMVRESGVEFC-------
#> Cfam_XRCC2 301 SNSLKKHTFVIGENGIEFC-------
#> Lafr_XRCC2 301 SNSLKKHFFIIGESGVEFC-------
#> Mdom_XRCC2 301 SNNLIKRLFSIRESGVHFC-------
#> Oana_XRCC2 301 SNRFAKRFFSIGEGGVQF--------
#> Ggal_XRCC2 301 KKHVAKRSFSIAECGVQFFQ------
#> Acar_XRCC2 301 NNSVIKRSFSILENGVQF--------
#> Xtro_XRCC2 301 TRNGVKRSFRIAESGVQFL-------
#> Drer_XRCC2 301 TKGVKKCSFCVVEDGIKFICDK----
#> Bflo_XRCC2 301 KPVEHKCEFTVADRGISFMGSLYNQC
#> Spur_XRCC2 301 SGNMYSCQFIIDQQGIRV--------
#> Nvec_XRCC2 301 KTESKSNRFLIEEHGVVFIS------
#> Tadh_XRCC2 301 SSIESAERFIVSEEGVVFI-------